Removed dependency to wxWidget's webview (not supported by Ubuntu provided wxWidgets version), removed integrated help (now opening directly the default browser)
